backend/requirements.txt
# Core API (install with: pip install -r requirements.txt)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# ML / model dependencies (on PyPI)
torch>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
mmengine>=0.9.0

# surgvlp is not on PyPI - install from your local clone or internal repo, e.g.:
#   pip install -e /path/to/surgvlp
# Then: pip install -r requirements-model.txt (if you have one)
